pyproject.toml
[project]
name = "post_process_jfsd"
version = "1.0.0"
description = "A python package to post process the jfsd data"
authors = [
  { name="Athanasios Machas", email="amachas@materials.uoc.gr" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "toml",
    "scipy",
    "matplotlib",
    "cmcrameri",
    "freud_analysis",

]

[project.scripts]
post_process_jfsd = "post_process_jfsd.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
